# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Domiciliaria Médica",
        "description": "Documentación oficial de la API para el sistema de atención domiciliaria",
        "contact": {
            "email": "admin@domiciliaria.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/ingresos": {
            "get": {
                "tags": [
                    "Ingresos"
                ],
                "summary": "Listar todos los ingresos",
                "operationId": "c4eb052a6babaefdab958f4de08724ee",
                "responses": {
                    "200": {
                        "description": "Lista de ingresos obtenida correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ingresos"
                ],
                "summary": "Registrar un nuevo ingreso",
                "operationId": "24d5b1100d776b7ac04fb2366f190487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ingreso",
                                    "id_paciente",
                                    "fecha_ingreso"
                                ],
                                "properties": {
                                    "ingreso": {
                                        "type": "integer",
                                        "example": 12345
                                    },
                                    "id_paciente": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "autorizacion": {
                                        "type": "string",
                                        "example": "AUT-001"
                                    },
                                    "fecha_ingreso": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-04-28 10:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ingreso registrado correctamente"
                    },
                    "400": {
                        "description": "Error en la solicitud"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/ingresos/paciente/{idPaciente}/autorizaciones": {
            "get": {
                "tags": [
                    "Ingresos"
                ],
                "summary": "Obtener todas las autorizaciones de un paciente",
                "operationId": "e4259b20c52acec63e5db95940360065",
                "parameters": [
                    {
                        "name": "idPaciente",
                        "in": "path",
                        "description": "ID del paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de autorizaciones obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "autorizacion": {
                                                        "type": "string",
                                                        "example": "AUT-2024-001"
                                                    },
                                                    "fecha_ingreso": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-04-28 10:00:00"
                                                    },
                                                    "ingreso": {
                                                        "type": "integer",
                                                        "example": 12345
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce tu token JWT con el formato **Bearer &lt;token&gt;**",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}